Create login & restrict access.
In order to restrict access using Devise, we had to update, change,
delete and add models, views, controllers. Other work was done to aid in
development (i.e. helpers, factories). In order to make this a 'base
commit' to build from, the code was cleaned, refactored and changed for
this purpose. This was a large amount of work
that needed to be changed and initially forgetting/neglecting to
properly use git led to this large commit.